refactor: extract faker methods, auth and user tests to separate files
